+++ tests/guest/t_fd_rw.c
@@ -86,6 +86,9 @@ int main(void) {
     T_ASSERT_ERRNO(pwritev(pp[1], &one, 1, 0), ESPIPE);
     T_ASSERT_ERRNO(lseek(pp[0], 0, SEEK_CUR), ESPIPE);
     T_ASSERT_ERRNO(lseek(pp[1], 0, SEEK_SET), ESPIPE);
+    T_ASSERT_ERRNO(ftruncate(pp[1], 0), EINVAL);
+    T_ASSERT_ERRNO(fsync(pp[1]), EINVAL);
+    T_ASSERT_ERRNO(fdatasync(pp[1]), EINVAL);
     T_ASSERT_OK(close(pp[0]));
     T_ASSERT_OK(close(pp[1]));
     fd = open("t_frw_x", O_RDWR | O_CREAT | O_TRUNC, 0600);